refactor: move WasmApplicationContext to separate file
