+++ src/Sitko.Core.Blazor.Wasm/WasmApplicationContext.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
+using Microsoft.Extensions.Configuration;
+using Sitko.Core.App;
+
+namespace Sitko.Core.Blazor.Wasm;
+
+public class WasmApplicationContext : BaseApplicationContext
+{
+    private readonly IWebAssemblyHostEnvironment environment;
+
+    public WasmApplicationContext(IConfiguration configuration, IWebAssemblyHostEnvironment environment)
+        : base(configuration) =>
+        this.environment = environment;
+
+    public override string EnvironmentName => environment.Environment;
+    public override bool IsDevelopment() => environment.IsDevelopment();
+
+    public override bool IsProduction() => environment.IsProduction();
+}
